Remove Season Interruption Insurance Core Component and Related Tests
- Deleted the `season-interruption-insurance-core.tsx` component and its associated test file to streamline the insurance module.
- Refactored the `season-interruption-insurance.tsx` component to incorporate necessary logic previously handled by the core component, ensuring continuity of functionality.
- Updated snapshot tests for the `season-interruption-insurance` component to reflect the changes in the rendered output and maintain test accuracy.
