+++ packages/webcomponents/src/components/insurance/season-interruption/season-interruption-insurance.tsx
@@ -6,6 +6,14 @@ import JustifiAnalytics from '../../../api/Analytics';
 import { checkPkgVersion } from '../../../utils/check-pkg-version';
 import { ComponentErrorEvent } from '../../../api/ComponentEvents';
 import { checkoutStore } from '../../../store/checkout.store';
+import { addAttribute, formatCurrency, processHTML, removeAttribute } from '../../../utils/utils';
+import { insuranceValues, insuranceErrors, validateInsuranceValues } from '../insurance-state';
+import { StyledHost, Header2 } from '../../../ui-components';
+import {
+  text,
+  textDanger
+} from '../../../styles/parts';
+import SeasonInterruptionInsuranceLoading from './season-interruption-insurance-loading';
 
 @Component({
   tag: 'justifi-season-interruption-insurance',
@@ -28,22 +36,35 @@ export class SeasonInterruptionInsurance {
   @State() getQuote: Function;
   @State() toggleCoverage: Function;
   @State() quote: any;
+  @State() isLoading: boolean = true;
+  @State() accepted: string | undefined;
 
+  @Event({ eventName: 'insurance-updated' }) insuranceUpdated: EventEmitter;
   @Event({ eventName: 'error-event' }) errorEvent: EventEmitter<ComponentErrorEvent>;
-
-  private seasonInterruptionCoreRef?: HTMLJustifiSeasonInterruptionInsuranceCoreElement;
 
   analytics: JustifiAnalytics;
 
   @Method()
   async validate(): Promise<{ isValid: boolean }> {
-    return this.seasonInterruptionCoreRef.validate();
+    return validateInsuranceValues();
+  }
+
+  processHTMLContent(html: string) {
+    return processHTML(html, [
+      (html) => removeAttribute(html, 'style'),
+      (html) => addAttribute(html, 'a', 'part', text),
+      (html) => addAttribute(html, 'p', 'part', text),
+    ]);
   }
 
   componentWillLoad() {
     checkPkgVersion();
     this.analytics = new JustifiAnalytics(this);
     this.initializeServiceMethods();
+
+    if (this.getQuote) {
+      this.fetchData();
+    }
   }
 
   disconnectedCallback() {
@@ -67,27 +88,112 @@ export class SeasonInterruptionInsurance {
     this.toggleCoverage = makeToggleCoverage(config);
   }
 
+  fetchData(): void {
+    this.isLoading = true;
+
+    this.getQuote({
+      payload: {
+        checkout_id: this.checkoutId || checkoutStore.checkoutId,
+        policy_type: "season_interruption",
+        primary_identity: {
+          state: this.primaryIdentityState,
+          email: this.primaryIdentityEmailAddress,
+          first_name: this.primaryIdentityFirstName,
+          last_name: this.primaryIdentityLastName,
+          postal_code: this.primaryIdentityPostalCode,
+          country: this.primaryIdentityCountry
+        },
+        policy_attributes: {
+          insurable_amount: this.policyAttributesInsurableAmount,
+          start_date: this.policyAttributesStartDate,
+          end_date: this.policyAttributesEndDate,
+          covered_identity: {
+            first_name: this.coveredIdentityFirstName,
+            last_name: this.coveredIdentityLastName
+          }
+        }
+      },
+      onSuccess: ({ quote }) => {
+        this.quote = quote;
+        this.quote.product.description = this.processHTMLContent(this.quote.product.description);
+        this.quote.product.legal_disclaimer = this.processHTMLContent(this.quote.product.legal_disclaimer);
+        insuranceValues[quote.policy_type] = quote.accepted;
+        this.isLoading = false;
+      },
+      onError: ({ error, code, severity }) => {
+        this.isLoading = false;
+        this.errorEvent.emit({
+          errorCode: code,
+          message: error,
+          severity,
+        });
+      }
+    });
+  }
+
+  onChangeHandler(_name: string, value: string) {
+    this.accepted = value;
+    insuranceErrors[this.quote.policy_type] = false;
+    this.toggleCoverage({
+      quoteId: this.quote.id,
+      payload: { accepted: this.accepted },
+      onSuccess: () => {
+        this.insuranceUpdated.emit();
+        insuranceValues[this.quote.policy_type] = this.accepted;
+      },
+      onError: ({ error, code, severity }) => {
+        this.errorEvent.emit({
+          errorCode: code,
+          message: error,
+          severity,
+        });
+      }
+    });
+  }
+
+  get error() {
+    return insuranceErrors[this.quote?.policy_type];
+  }
 
   render() {
+    if (this.isLoading) return (
+      <StyledHost>
+        <SeasonInterruptionInsuranceLoading />
+      </StyledHost>
+    );
+
     return (
-      <justifi-season-interruption-insurance-core
-        ref={(el) => (this.seasonInterruptionCoreRef = el)}
-        checkout-id={this.checkoutId || checkoutStore.checkoutId}
-        getQuote={this.getQuote}
-        toggleCoverage={this.toggleCoverage}
-        primary-identity-first-name={this.primaryIdentityFirstName}
-        primary-identity-last-name={this.primaryIdentityLastName}
-        primary-identity-state={this.primaryIdentityState}
-        primary-identity-postal-code={this.primaryIdentityPostalCode}
-        primary-identity-country={this.primaryIdentityCountry}
-        primary-identity-email-address={this.primaryIdentityEmailAddress}
-        policy-attributes-insurable-amount={this.policyAttributesInsurableAmount}
-        policy-attributes-start-date={this.policyAttributesStartDate}
-        policy-attributes-end-date={this.policyAttributesEndDate}
-        covered-identity-first-name={this.coveredIdentityFirstName}
-        covered-identity-last-name={this.coveredIdentityLastName}
-      >
-      </justifi-season-interruption-insurance-core>
+      <StyledHost>
+        <div>
+          <Header2 text={this.quote?.product.title} class="fs-5 fw-bold pb-3" />
+          <small part={text} innerHTML={this.quote?.product.description}></small>
+          <form-control-radio
+            label={`Accept coverage for ${formatCurrency(this.quote?.total_cents)}`}
+            name="opt-in"
+            value={'true'}
+            checked={this.accepted === 'true'}
+            inputHandler={this.onChangeHandler.bind(this)}
+            // don't wanna show error message, but need to show the red border
+            errorText={this.error ? ' ' : undefined}
+          />
+          <form-control-radio
+            label="Decline coverage"
+            name="opt-in"
+            value={'false'}
+            checked={this.accepted === 'false'}
+            inputHandler={this.onChangeHandler.bind(this)}
+            errorText={this.error ? ' ' : undefined}
+          />
+          <div
+            class="invalid-feedback"
+            style={{ display: this.error ? 'block' : 'none' }}
+            part={textDanger}
+          >
+            Please select an option
+          </div>
+          <small part={text} innerHTML={this.quote?.product.legal_disclaimer}></small>
+        </div>
+      </StyledHost>
     );
   }
 }
